Move real module to wrf subpackage
