+++ erftools/preprocessing/__init__.py
@@ -1,4 +1,3 @@
 from .wrf_inputs import WRFInputDeck
 from .grids import LambertConformalGrid
 from .hrrr import NativeHRRR, hrrr_projection
-from .real import get_zlevels_auto
