+++ erftools/preprocessing/hrrr.py
@@ -8,7 +8,7 @@ from herbie import Herbie
 from ..constants import R_d, R_v, Cp_d, Cp_v, CONST_GRAV, p_0
 from ..EOS import getPgivenRTh, getThgivenRandT, getThgivenPandT
 from ..utils import get_hi_faces, get_lo_faces, get_w_from_omega
-from .real import RealInit
+from ..wrf.real import RealInit
 
 
 hrrr_projection = ccrs.LambertConformal(
