+++ erftools/wrf/__init__.py
@@ -1,0 +1,2 @@
+from .real import get_zlevels_auto
+
